add image generation feature

+++ src/services/event.service.ts
@@ -1,4 +1,5 @@
 import { Event, Prisma } from "../../generated/prisma/client";
+import { prisma } from "../lib/prisma";
 
 import {
   countWeddingEvents,
@@ -9,6 +10,7 @@ import {
   getAllWeddingEvents,
   updateWeddingEventById,
 } from "../repositories/event.repository";
+import { createGuestEventInviteFormat } from "../repositories/eventInviteFormat.repository";
 import { ApiError } from "../utils/apiError.util";
 
 export const getAllWeddingEventsService = async (
@@ -39,36 +41,57 @@ export const getAllWeddingEventsService = async (
 export const addNewWeddingEventService = async (
   data: Prisma.EventUncheckedCreateInput,
 ): Promise<Event> => {
-  const event = await createEvent({
-    wedding_id: data.wedding_id,
-    title: data.title,
-    description: data.description,
-    date: new Date(data.date),
-    time: data.time,
-    venue: data.venue,
-    address: data.address,
-    city: data.city,
+  return prisma.$transaction(async (tx) => {
+    const event = await createEvent(
+      {
+        wedding_id: data.wedding_id,
+        title: data.title,
+        description: data.description,
+        event_side: data.event_side,
+        date: new Date(data.date),
+        time: data.time,
+        venue: data.venue,
+        address: data.address,
+        city: data.city,
+      },
+      tx,
+    );
+
+    if (!event) throw new ApiError(400, "Failed to create event");
+
+    const eventInviteFormat = await createGuestEventInviteFormat(
+      { event_id: event.id },
+      tx,
+    );
+
+    if (!eventInviteFormat)
+      throw new ApiError(400, "Failed to create event invite format");
+
+    return event;
   });
-
-  if (!event) throw new ApiError(400, "Failed to create event");
-
-  return event;
 };
 
 export const verifyWeddingEventOwnershipService = async (
   eventId: string,
   userId: string,
-): Promise<Event> => {
+): Promise<Event | null> => {
   const event = await findEventByIdAndUserId(eventId, userId);
 
-  if (!event) throw new ApiError(404, "Event not found");
-
-  return event
+  return event;
 };
 
 export const getWeddingEventService = async (
   eventId: string,
+  userId: string,
 ): Promise<Event> => {
+  const ownershipEvent = await verifyWeddingEventOwnershipService(
+    eventId,
+    userId,
+  );
+
+  if (!ownershipEvent)
+    throw new ApiError(400, "Invalid Event or Event Not Found");
+
   const event = await findEventById(eventId);
 
   if (!event) throw new ApiError(404, "Event not found");
@@ -78,8 +101,14 @@ export const getWeddingEventService = async (
 
 export const editWeddingEventService = async (
   id: string,
+  userId: string,
   payload: Prisma.EventUpdateInput,
 ): Promise<Event> => {
+  const ownershipEvent = await verifyWeddingEventOwnershipService(id, userId);
+
+  if (!ownershipEvent)
+    throw new ApiError(400, "Invalid Event or Event Not Found");
+
   if (payload.date) {
     payload.date = new Date(payload.date as string);
   }
@@ -93,7 +122,16 @@ export const editWeddingEventService = async (
 
 export const deleteWeddingEventService = async (
   eventId: string,
+  userId: string,
 ): Promise<void> => {
+  const ownershipEvent = await verifyWeddingEventOwnershipService(
+    eventId,
+    userId,
+  );
+
+  if (!ownershipEvent)
+    throw new ApiError(400, "Invalid Event or Event Not Found");
+
   const event = await deleteWeddingEventById(eventId);
 
   if (!event) throw new ApiError(400, "Failed to delete event");
